moved Loop-matching of `+` below the non-loop matching one.
The way that the matching code is set up, the loop-version was
catching cases that were *not* loop starts, and thus injecting `M {C}`
into the middle of paths that really wanted an `L {C}` instead.

I would probably be better off classifying the matching of `Loop` in
non-loop cases as a bug, and fixing it, rather than trying to keep in
mind an additional order-dependence in the format table. But one thing
at a time.