Debugging hack: Add loud rules and inspection sigils `%` and `$`.
If we encounter `%` "near a path", we will start printing out the
handling of all the "loud" rules (which are currently encoded by
wrapping each such rule in `Loud(..)`).

Encountering `$` in a similar manner turns off the instrumentation.

Note that only loud rules are instrumented; I found that the amount of
output generated if you ask to see all rules was unmanagable.  (And
similarly, asking to see all uses of any one rule on all paths under
consideration was also unmanagable.)