removed hack that treated `->` specially.
if you want to make a path not get read as a path, then you can use
the cancellation character ☠.

(and what if you don't want the cancellation character to end up as a
text element of its own in the diagram? One workaround may be to add
an attribute *that* element, making it invisible by e.g. setting its
color. Another option is that I may add an optional mode that simply
removes all occurrences of the cancellation character prior to
scanning for text. That might be the best option here.)