Render circle joints via new `/o` suffix.
This seems to handle straight lines going through a `o` in the manner
I think I expected. But as soon as you squeeze the joint so that it
has less than 180 degrees, the circle starts to shrink, which was not
my original intention.

(However, to be honest this accidental semantics may actually be the
better option, since I could imagine a big circle looking out of a
place when the angle gets really sharp.)