Attribute parsing support. Enabled front-end based rule set by default.
To this end, I changed the language slighty: I added a terminator
(`;`) to each rule, and I prefix the attribute with the keyword
`attrs` in each rule that uses them. (My intention is to allow for
future extensions to the rule specification language; e.g.  I have
been thinking recently of allowing "decorator" paths that do not
contribute to the main one, but will be added to the whole path after
the main one is rendered.