Add options struct for Scene construction.
Allow user to override whether we interpret `[...]` neighboring either
a path or a text object as a markdown label. The reason for this is
that I have discovered rendering annotated code is a non-starter if
`a[i]` is interpreted as a text object "a" with id `i`. So now you can
turn that off.

(I am considering making this more fine-grained, e.g. where
text-object labels are based solely on neighbors on the previous or
next line, but not neighbors on the same line.)